+++ app/src/main/java/com/mataku/scrobscrob/app/ui/navigation/NavigationGraph.kt
@@ -8,10 +8,10 @@ import androidx.navigation.compose.dialog
 import androidx.navigation.navArgument
 import com.mataku.scrobscrob.album.ui.screen.TopAlbumsScreen
 import com.mataku.scrobscrob.app.ui.molecule.LogoutConfirmationDialog
-import com.mataku.scrobscrob.app.ui.molecule.SunsetBottomNavItem
-import com.mataku.scrobscrob.app.ui.screen.LoginScreen
 import com.mataku.scrobscrob.app.ui.screen.ScrobbleScreen
 import com.mataku.scrobscrob.artist.ui.screen.TopArtistsScreen
+import com.mataku.scrobscrob.login.ui.screen.LoginScreen
+import com.mataku.scrobscrob.ui_common.SunsetBottomNavItem
 import com.mataku.scrobscrob.ui_common.template.WebViewScreen
 
 // @OptIn(ExperimentalAnimationApi::class)
