+++ ui_common/src/main/res/drawable/ic_album_black_24px.xml
