+++ ui_common/src/main/res/drawable/ic_last_fm_logo.xml
